+++ sw.js
@@ -13,7 +13,7 @@
  */
 'use strict';
 
-var VERSION = 'v53';
+var VERSION = 'v54';
 var RUNTIME_CACHE_VERSION = 'v2';
 var SHELL_CACHE = 'stormscope-shell-' + VERSION;
 var TILE_CACHE = 'stormscope-tiles-' + RUNTIME_CACHE_VERSION;
@@ -42,6 +42,7 @@ var SHELL_ASSETS = [
   './js/multi-camera.js',
   './js/context-layers.js',
   './js/diagnostics.js',
+  './js/spatial-query.js',
   './js/app.js',
   './vendor/leaflet/leaflet.css',
   './vendor/leaflet/leaflet.js',